Fix seeking in encoded files. So the encoder returns Annex B format H.264 data where all NALUs have a prefix of 0x00000001 (4-byte sequence).
To properly mux this into MKV/MP4s, the SPS/PPS data must be stored in
the extradata/CodecPrivate area in the AVCC format. The SPS/PPS data
is not stored in the stream itself (although modern versions of libav
can handle this). Furthermore, the 0x00000001 prefix must then be
replaced by the length of the NALU in big endian format. Note that this
length is the length of the returned buffer minus 4 bytes, since we
do not include the length of the prefx.

Also, because of this requirement, the header can only be written out
once we have our set of SPS and PPS data. Hence deferring the
avformat_write_header operation until this point in time.