Fixup NEON pixel count parameter (still untested) + testing fragmentation setting.
It seems that when OMX_IndexConfigMinimiseFragmentation is enabled,
NALUs do tend to come out in whole pieces and not split across buffers.

However, one quirk is that the SPS and PPS are received in one buffer,
separated by the start sequence, even though we indicate that we want
one NALU per buffer. The code has been reworked to not segfault if
we don't have an SPS/PPS combo before any other NALU is received.

Brief testing suggests that this concatenation of NALUs into one buffer
does not occur elsewhere.

Brief testing however also suggests that enabling this is slightly slower
than just storing it in our own buffer. More testing required to confirm
this.